requirements.txt
streamlit
pandas
plotly.express
matplotlib.plotly
matplotlib
